Updated auth urls

+++ src/web/pages/static/assets/components/user-list.jsx
@@ -4,12 +4,14 @@ define([
         'react',
         'react-dom',
         'moment',
-        'event-manager',
+        'global-events',
+        'event',
         'jquery'],
     function (React,
               ReactDom,
               Moment,
-              EventManager,
+              GlobalEvents,
+              Event,
               $) {
 
         var url = "/api/v1/users";
@@ -23,7 +25,9 @@ define([
             },
 
             componentDidMount: function () {
-                EventManager.on('user-created, user-updated, user-deleted', this.retrieveDataAndUpdateList.bind(this));
+                GlobalEvents.userCreated.on(this.retrieveDataAndUpdateList.bind(this));
+                GlobalEvents.userUpdated.on(this.retrieveDataAndUpdateList.bind(this));
+                GlobalEvents.userDeleted.on(this.retrieveDataAndUpdateList.bind(this));
                 this.retrieveDataAndUpdateList();
             },
 
@@ -56,19 +60,30 @@ define([
                                 <th>Name</th>
                                 <th>Email</th>
                                 <th></th>
+                                <th></th>
                                 <th className="m-usl-hidden-cell">Created</th>
                                 <th className="m-usl-hidden-cell">Updated</th>
                             </tr>
                             </thead>
                             <tbody>
                             {this.state.users.map(function (user, index) {
-                                return <UserListItem key={user._id} id={user._id} user={user} number={index+1}/>;
+                                return <UserListItem key={user._id} user={user} number={index+1}/>;
                             })}
                             </tbody>
                         </table>
+
+                        <div>
+                            <a className="m-usl-add-button" onClick={this.onAddButtonClicked}><i className="fa fa-plus-square-o m-usl-add-icon"></i> Add</a>
+                        </div>
                     </div>
                 );
+            },
+
+            onAddButtonClicked: function (e) {
+                e.preventDefault();
+                exports.onAddButtonClicked.fire();
             }
+
         });
 
         var UserListItem = React.createClass({
@@ -83,38 +98,58 @@ define([
                         <th>{this.props.number}</th>
                         <td>{this.props.user.display_name}</td>
                         <td>{this.props.user.email}</td>
-                        <td className="m-usl-center"><i className="fa fa-trash-o m-usl-trash-icon" onClick={this.onTrashIconClicked}></i></td>
+                        <td className="m-usl-center"><i title="Edit" className="fa fa-pencil-square-o m-usl-edit-icon" onClick={this.onEditButtonClicked}></i></td>
+                        <td className="m-usl-center"><i title="Delete" className="fa fa-trash-o m-usl-delete-icon" onClick={this.onDeleteButtonClicked}></i></td>
                         <td className="m-usl-hidden-cell">{created}</td>
                         <td className="m-usl-hidden-cell">{updated}</td>
                     </tr>
                 );
             },
 
-            onTrashIconClicked: function (e) {
+            onDeleteButtonClicked: function (e) {
                 e.preventDefault();
+                exports.onDeleteButtonClicked.fire(this.props.user);
 
-                $.ajax({
-                    url: url + "/" + this.props.id,
-                    dataType: 'json',
-                    cache: false,
-                    method: 'delete'
-                })
-                    .then(function () {
-                        EventManager.fire('user-deleted');
-                    }.bind(this))
+                /*
+                var message = 'Do you want to delete "' + this.props.user.display_name + '"?';
 
-                    .fail(function (xhr, status, err) {
-                        console.error(url, status, err.toString());
-                    }.bind(this));
+                if (confirm(message)) {
+                    $.ajax({
+                        url: url + "/" + this.props.user._id,
+                        dataType: 'json',
+                        cache: false,
+                        method: 'delete'
+                    })
+                        .then(function () {
+                            GlobalEvents.userDeleted.fire();
+                        }.bind(this))
+
+                        .fail(function (xhr, status, err) {
+                            console.error(url, status, err.toString());
+                        }.bind(this));
+                }
+                */
+            },
+
+            onEditButtonClicked: function (e) {
+                e.preventDefault();
+                exports.onEditButtonClicked.fire(this.props.user);
             }
 
         });
 
-        return {
+
+        var exports = {
+
+            onAddButtonClicked: new Event(),
+            onEditButtonClicked: new Event(),
+            onDeleteButtonClicked: new Event(),
+
             render: function (container) {
                 ReactDom.render(<UserList />, container);
             }
         };
 
+        return exports;
     });
